+++ noodles-csi/src/writer.rs
@@ -104,15 +104,24 @@ where
     let format = i32::from(header.format());
     writer.write_i32::<LittleEndian>(format)?;
 
-    let col_seq = i32::try_from(header.reference_sequence_name_index())
+    let reference_sequence_name_index = header
+        .reference_sequence_name_index()
+        .checked_add(1)
+        .expect("attempt to add with overflow");
+    let col_seq = i32::try_from(reference_sequence_name_index)
         .map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))?;
     writer.write_i32::<LittleEndian>(col_seq)?;
 
-    let col_beg = i32::try_from(header.start_position_index())
+    let start_position_index = header
+        .start_position_index()
+        .checked_add(1)
+        .expect("attempt to add with overflow");
+    let col_beg = i32::try_from(start_position_index)
         .map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))?;
     writer.write_i32::<LittleEndian>(col_beg)?;
 
-    let col_end = header.end_position_index().map_or(Ok(0), |i| {
+    let col_end = header.end_position_index().map_or(Ok(0), |mut i| {
+        i = i.checked_add(1).expect("attempt to add with overflow");
         i32::try_from(i).map_err(|e| io::Error::new(io::ErrorKind::InvalidInput, e))
     })?;
     writer.write_i32::<LittleEndian>(col_end)?;
@@ -274,6 +283,31 @@ where
 #[cfg(test)]
 mod tests {
     use super::*;
+
+    #[test]
+    fn test_write_header() -> io::Result<()> {
+        let header = crate::index::header::Builder::vcf()
+            .set_reference_sequence_names([String::from("sq0")].into_iter().collect())
+            .build();
+
+        let mut buf = Vec::new();
+        write_header(&mut buf, &header)?;
+
+        let expected = [
+            0x02, 0x00, 0x00, 0x00, // format = 2 (VCF)
+            0x01, 0x00, 0x00, 0x00, // col_seq = 1 (1-based)
+            0x02, 0x00, 0x00, 0x00, // col_beg = 2 (1-based)
+            0x00, 0x00, 0x00, 0x00, // col_end = None (1-based)
+            0x23, 0x00, 0x00, 0x00, // meta = '#'
+            0x00, 0x00, 0x00, 0x00, // skip = 0
+            0x04, 0x00, 0x00, 0x00, // l_nm = 4
+            b's', b'q', b'0', 0x00, // names = ["sq0"]
+        ];
+
+        assert_eq!(buf, expected);
+
+        Ok(())
+    }
 
     #[test]
     fn test_write_metadata() -> io::Result<()> {